requirements.txt
streamlit
pyswisseph
pandas
plotly